fix: Autoformat should not always be run on save
This will need to be fine-tuned because it should really use the settings in the repository if they are available, but for now this will have to do. It was automatically formatting every file on save, which complicated issues when I tried to contribute code without forcing arbitrary style along with the pull request.